Refactor

+++ source/Services/Interfaces/IDividendService.cs
@@ -9,9 +9,9 @@ namespace api.Services.Interfaces
     public interface IDividendService
     {
         List<Dividend> Get();
-        Dividend Get(int id);
+        Dividend Get(long id);
         bool Create(Dividend dividend);
         bool Update(Dividend dividend);
-        bool Delete(int id);
+        bool Delete(long id);
     }
 }
